Add proposal locking and unlocking endpoints to admin API
